refactor: break circular dependency between profile-service and builtin-tools
Addresses PR review comment about circular module dependency.

Created a new dependency-free module 'builtin-tool-definitions.ts' that contains:
- BUILTIN_SERVER_NAME constant
- builtinToolDefinitions array (static tool definitions)
- getBuiltinToolNames() helper function

This breaks the circular dependency:
- profile-service.ts now imports from builtin-tool-definitions.ts (no cycle)
- builtin-tools.ts re-exports from builtin-tool-definitions.ts for backward compatibility
- builtin-tools.ts still imports profileService for tool handlers (valid direction)

All TypeScript compilation and 36 tests pass.
